+++ server/ButtonsVital.js
@@ -1,8 +1,10 @@
 class ButtonsVital {
-  constructor(id, batteryLevel, createdAt, button) {
+  constructor(id, batteryLevel, createdAt, snr, rssi, button) {
     this.id = id
     this.batteryLevel = batteryLevel
     this.createdAt = createdAt
+    this.snr = snr
+    this.rssi = rssi
     this.button = button
   }
 }